requirements.txt
Flask==3.0.0
gunicorn==21.2.0
gspread==6.0.0
google-auth==2.27.0
google-api-python-client==2.111.0
